Add new prune mode
