+++ Source/Core/CoreOptions.cs
@@ -3,7 +3,11 @@ using System.IO;
 
 namespace Microsoft.Boogie
 {
-
+  public enum PruneMode
+  {
+    Not, AllRevealed, ExplicitReveal 
+  }
+  
   /// <summary>
   /// Boogie command-line options (other tools can subclass this class in order to support a
   /// superset of Boogie's options).
@@ -74,7 +78,7 @@ namespace Microsoft.Boogie
     int VcsCores { get; }
     int RandomizeVcIterations { get; }
     List<string> ProverOptions { get; }
-    bool Prune { get; }
+    PruneMode Prune { get; }
     bool RunDiagnosticsOnTimeout { get; }
     string ProverLogFilePath { get; set; }
     bool ProverLogFileAppend { get; }
