+++ Source/VCGeneration/CheckerPool.cs
@@ -65,7 +65,7 @@ namespace VC
 
     void PrepareChecker(Program program, Split? split, Checker checker)
     {
-      if (checker.WillingToHandle(program) && (split == null || checker.SolverOptions.RandomSeed == split.RandomSeed && !Options.Prune))
+      if (checker.WillingToHandle(program) && (split == null || checker.SolverOptions.RandomSeed == split.RandomSeed && Options.Prune == PruneMode.Not))
       {
         checker.GetReady();
         return;
