+++ Source/VCGeneration/Prune/Prune.cs
@@ -4,11 +4,12 @@ using Microsoft.Boogie.GraphUtil;
 
 namespace Microsoft.Boogie
 {
+  
   public class Prune {
 
     public static Dictionary<object, List<object>> ComputeDeclarationDependencies(VCGenOptions options, Program program)
     {
-      if (!options.Prune)
+      if (options.Prune == PruneMode.Not)
       {
         return null;
       }
@@ -39,7 +40,7 @@ namespace Microsoft.Boogie
           var targets = edges.GetOrCreate(source, () => new());
           targets.Add(node.declaration);
         }
-        foreach (var outgoingSingle in node.outgoing) {
+        foreach (var outgoingSingle in node.Outgoing) {
           var targets = edges.GetOrCreate(node.declaration, () => new());
           targets.Add(outgoingSingle);
         }
@@ -55,18 +56,34 @@ namespace Microsoft.Boogie
      */
     public static IEnumerable<Declaration> GetLiveDeclarations(VCGenOptions options, Program program, List<Block> blocks)
     {
-      if (program.DeclarationDependencies == null || blocks == null || !options.Prune)
+      if (program.DeclarationDependencies == null || blocks == null || options.Prune == PruneMode.Not)
       {
         return program.TopLevelDeclarations;
       }
 
-      BlocksVisitor blocksNode = new BlocksVisitor(blocks);
-      blocksNode.Blocks.ForEach(blk => blocksNode.Visit(blk));
+      PruneBlocksVisitor blocksVisitor = new PruneBlocksVisitor();
+      foreach (var block in blocks)
+      {
+        blocksVisitor.Visit(block);
+      }
+
+      var revealedOutgoing = options.Prune == PruneMode.AllRevealed
+        ? blocksVisitor.Outgoing
+        : blocksVisitor.Outgoing.Where(d =>
+          DeclarationFilter(d)); 
 
       var keepRoots = program.TopLevelDeclarations.Where(d => QKeyValue.FindBoolAttribute(d.Attributes, "keep"));
-      var reachableDeclarations = GraphAlgorithms.FindReachableNodesInGraphWithMergeNodes(program.DeclarationDependencies, blocksNode.outgoing.Concat(keepRoots).ToHashSet()).ToHashSet();
+      var reachableDeclarations = GraphAlgorithms.FindReachableNodesInGraphWithMergeNodes(program.DeclarationDependencies, 
+        revealedOutgoing.Concat(keepRoots).ToHashSet(),
+        node => node is not Function function || blocksVisitor.RevealedFunctions.Contains(function))
+        .OfType<Declaration>().Where(d => DeclarationFilter(d)).ToHashSet();
       return program.TopLevelDeclarations.Where(d => 
         d is not Constant && d is not Axiom && d is not Function || reachableDeclarations.Contains(d));
+
+      bool DeclarationFilter(Declaration d)
+      {
+        return d is not Function function || blocksVisitor.RevealedFunctions.Contains(function);
+      }
     }
   }
 }
