+++ Test/test0/AssumeFalseSplit/AssumeFalseSplit.bpl
@@ -4,6 +4,7 @@
 // RUN: %diff "%s.expect" "%t"
 // RUN: %diff %S/Foo.split.0.bpl.expect ./Output/Foo.split.0.bpl
 // RUN: %diff %S/Foo.split.1.bpl.expect ./Output/Foo.split.1.bpl
+// RUN: %diff %S/Foo.split.2.bpl.expect ./Output/Foo.split.2.bpl
 
 procedure Foo()
 {
@@ -12,7 +13,7 @@ procedure Foo()
     assume false; 
   } 
   assume 2 == 1 + 1; 
-  assert 2 == 2;
+  assert {:split_here} 2 == 2;
   assert {:split_here} 3 == 2 + 1;
   assert 1 == 1;
 }
