+++ Test/test0/AssumeFalseSplit/Foo.split.2.bpl.expect
@@ -6,9 +6,9 @@ implementation Foo()
 
   anon3_Else:
     assume 2 == 1 + 1;
-    assert {:split_here} 2 == 2;
-    assume 3 == 2 + 1;
-    assume 1 == 1;
+    assume 2 == 2;
+    assert {:split_here} 3 == 2 + 1;
+    assert 1 == 1;
     goto ;
 }
